修复统计出的总利润只能是证书的问题

+++ src/main/java/com/example/demo/domain/dto/StatisticsDTO.java
@@ -10,8 +10,8 @@ import lombok.Data;
 @Data
 public class StatisticsDTO {
 
-    private Integer totalSales;
-    private Integer totalProfit;
+    private Float totalSales;
+    private Float totalProfit;
     private Integer totalNotSold;
     private Integer totalSold;
